avoid warnings on win32
git-svn-id: https://svn.osgeo.org/gdal/trunk@21848 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/pcidsk/pcidskdataset2.cpp
@@ -443,9 +443,9 @@ bool PCIDSK2Band::CheckForColorTable()
             // we have an entry - apply to the color table.
             GDALColorEntry sEntry;
 
-            sEntry.c1 = nRed;
-            sEntry.c2 = nGreen;
-            sEntry.c3 = nBlue;
+            sEntry.c1 = (short) nRed;
+            sEntry.c2 = (short) nGreen;
+            sEntry.c3 = (short) nBlue;
             sEntry.c4 = 255;
 
             if( poColorTable == NULL )
